Module fractionalizer::state

Structs

Config

UploadedFtkn

the code_id and code hash of the ftoken contract code that has been uploaded

Constants

CONFIG_KEY

FTKN_INDEX

FTOKEN_CONTR_FRAC

PENDING_REG

UPLOADED_FTKN

Functions

config_r

config_w

ftkn_id_hash_r

ftkn_id_hash_w

stores the code_id and code hash of the ftoken contract code that has been uploaded

ftkn_idx_r

ftkn_idx_w

index of next ftoken contract to be created

ftoken_instance_r

ftoken_instance_w

FtokenContr storage: stores information on the ftokens that fractionalizer contract has created

pending_reg_r

pending_reg_w

pending info (HumanAddr of depositor) of ftoken contract to be registered, so fractionalizer can verify the callback from ftoken contract. Info should not last beyond a transaction